requirements.txt
websocket-client
python-Levenshtein
streamlit
pandas